+++ src/extension.ts
@@ -6,13 +6,12 @@ import { EulerLensProvider } from "./lens";
 import { EulerResourceProvider } from "./resource";
 
 async function searchProblem() {
-    const problemNumber = await vscode.window
-        .showInputBox({
-            placeHolder: "Enter Problem #",
-        });
+    const problemNumber = await vscode.window.showInputBox({
+        placeHolder: "Enter Problem #",
+    });
     if (problemNumber) {
         vscode.commands.executeCommand(Command.Show, problemNumber);
-    };
+    }
 }
 
 export function activate(context: vscode.ExtensionContext) {
@@ -20,19 +19,19 @@ export function activate(context: vscode.ExtensionContext) {
     const dataService = new ProblemDataService(context);
     const viewProvider = new ProblemViewProvider(context, dataService);
     const resourceProvider = new EulerResourceProvider(context);
-    const treeProvider = new ProblemTreeDataProvider(dataService);
+    const treeProvider = new ProblemTreeDataProvider(dataService, context);
     context.subscriptions.push(
         vscode.workspace.registerTextDocumentContentProvider("resource", resourceProvider),
         vscode.commands.registerCommand(Command.Search, searchProblem),
         vscode.commands.registerCommand(Command.Show, viewProvider.showProblem, viewProvider),
-        vscode.commands.registerCommand(Command.Refresh, treeProvider.refresh, treeProvider),
-        vscode.commands.registerCommand(Command.Clear, dataService.clearProblemInfo, dataService),
-        vscode.commands.registerCommand(Command.Load, dataService.updateProblemInfo, dataService),
+        vscode.commands.registerCommand(Command.Refresh, dataService.refreshMetadata, dataService),
+        vscode.commands.registerCommand(Command.Clear, dataService.clearMetadata, dataService),
     );
 
     vscode.window.registerTreeDataProvider("projecteuler.problemView", treeProvider);
     vscode.window.registerFileDecorationProvider(new ProblemTreeDecorationProvider());
+    dataService.init();
 }
 
 // This method is called when your extension is deactivated
-export function deactivate() { }
+export function deactivate() {}